requirements.txt
requests
selenium
webdriver-manager
torch
torchvision
transformers
Pillow
python-dotenv
